requirements.txt
python-dotenv==1.0.1
psycopg[binary,pool]==3.1.18
